SERVER-9763 Remove BSONObjBuilder append unsigned method
GitOrigin-RevId: 5ade901fef94f245be6780ec04c3b0b4af1442a5

+++ src/mongo/db/modules/enterprise/src/audit/audit_event.cpp
@@ -35,12 +35,12 @@ void AuditEvent::generateBSON() const {
     {
         BSONObjBuilder localIpBuilder(builder.subobjStart("local"));
         builder.append("ip", getLocalAddr().getAddr());
-        builder.append("port", getLocalAddr().getPort());
+        builder.append("port", static_cast<int>(getLocalAddr().getPort()));
     }
     {
         BSONObjBuilder remoteIpBuilder(builder.subobjStart("remote"));
         builder.append("ip", getRemoteAddr().getAddr());
-        builder.append("port", getRemoteAddr().getPort());
+        builder.append("port", static_cast<int>(getRemoteAddr().getPort()));
     }
     {
         UserNameIterator userNames = getImpersonatedUserNames();
